CREATE TABLE with column collation as latin1_general_90_bin2 fails after MVU
Collation latin1_general_90_bin2, translated to bbf_unicode_bin2,
has collencoding = -1 on a freshly installed server but it gets
changed to databaseCollation (collencoding = PG_UTF8) after MVU.
While in lookup_collation function, we search collation tuple
only with encoding = -1 and fail to find the tuple. That's why
CREATE TABLE with this collation fails with the error that
"collation %s for encoding UTF8 does not exists".

Fixed this by adding a second search with given encoding if
we fail to find the collation with encoding = -1.

Signed-off-by: Rishabh Tanwar <ritanwar@amazon.com>