Clean up some residual confusion between OIDs and RelFileNumbers.
Commit b0a55e43299c4ea2a9a8c757f9c26352407d0ccc missed a few places
where we are referring to the number used as a part of the relation
filename as an "OID". We now want to call that a "RelFileNumber".

Some of these places actually made it sound like the OID in question
is pg_class.oid rather than pg_class.relfilenode, which is especially
good to clean up.

Dilip Kumar with some editing by me.

(cherry picked from commit 851f4cc75cdd8c831f1baa9a7abf8c8248b65890)